refactor: extract admin blacklist handlers to layered architecture
- Create api/admin/blacklist.rs
- Create repositories/blacklist_repository.rs
- Remove 5 SQL functions from db/queries.rs
- Replace internal get_all_blacklist() call with inline SQL query
- Remove 3 handlers from router.rs
