+++ src/api/admin/mod.rs
@@ -1,1 +1,2 @@
 pub mod api_keys;
+pub mod blacklist;
