+++ src/router.rs
@@ -2506,247 +2506,7 @@ pub async fn handle_admin_sync_link_kv(req: Request, ctx: RouteContext<()>) -> R
     }))
 }
 
-#[utoipa::path(
-    post,
-    path = "/api/admin/blacklist",
-    tag = "Admin",
-    summary = "Block a destination",
-    responses(
-        (status = 200, description = "Destination blocked"),
-        (status = 400, description = "Invalid request"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_block_destination(
-    mut req: Request,
-    ctx: RouteContext<()>,
-) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    // Parse request body
-    let body: serde_json::Value = match req.json().await {
-        Ok(body) => body,
-        Err(e) => return Response::error(format!("Invalid JSON: {}", e), 400),
-    };
-
-    let destination = match body.get("destination").and_then(|d| d.as_str()) {
-        Some(d) => d.to_string(),
-        None => return Response::error("Missing 'destination' field", 400),
-    };
-
-    let match_type = match body.get("match_type").and_then(|m| m.as_str()) {
-        Some(m) if m == "exact" || m == "domain" => m.to_string(),
-        Some(_) => return Response::error("Invalid match_type. Must be 'exact' or 'domain'", 400),
-        None => "exact".to_string(), // Default to exact
-    };
-
-    // Normalize the destination URL for exact matches to ensure consistency
-    let normalized_destination = if match_type == "exact" {
-        match crate::utils::normalize_url_for_blacklist(&destination) {
-            Ok(url) => url,
-            Err(e) => {
-                console_log!(
-                    "{}",
-                    serde_json::json!({
-                        "event": "url_normalize_failed",
-                        "url": destination,
-                        "error": e.to_string(),
-                        "level": "warn"
-                    })
-                );
-                // Fall back to original URL if normalization fails
-                destination.clone()
-            }
-        }
-    } else {
-        // For domain matches, extract just the domain/hostname
-        match url::Url::parse(&destination) {
-            Ok(url) => url.host_str().unwrap_or(&destination).to_string(),
-            Err(_) => {
-                console_log!(
-                    "{}",
-                    serde_json::json!({
-                        "event": "domain_extract_failed",
-                        "url": destination,
-                        "level": "warn"
-                    })
-                );
-                // If URL parsing fails, try simple extraction
-                let without_protocol = destination
-                    .trim_start_matches("http://")
-                    .trim_start_matches("https://");
-                let without_path = without_protocol
-                    .split('/')
-                    .next()
-                    .unwrap_or(without_protocol);
-                without_path.to_string()
-            }
-        }
-    };
-
-    let reason = match body.get("reason").and_then(|r| r.as_str()) {
-        Some(r) => r.to_string(),
-        None => return Response::error("Missing 'reason' field", 400),
-    };
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Check if destination is already blacklisted
-    if db::is_destination_already_blacklisted(&db, &normalized_destination, &match_type).await? {
-        return Response::from_json(&serde_json::json!({
-            "success": false,
-            "message": "Destination is already blocked",
-            "already_blocked": true
-        }));
-    }
-
-    // Add to blacklist
-    db::add_to_blacklist(
-        &db,
-        &normalized_destination,
-        &match_type,
-        &reason,
-        &user_ctx.user_id,
-    )
-    .await?;
-
-    let candidate_links = db::get_links_for_blacklist_scan(&db).await?;
-
-    let mut blocked_links = Vec::new();
-    for link in candidate_links {
-        if db::is_destination_blacklisted(&db, &link.destination_url).await? {
-            blocked_links.push(link);
-        }
-    }
-
-    let blocked_count = blocked_links.len() as i64;
-
-    // Auto-resolve all pending reports for the blocked links
-    if blocked_count > 0 {
-        let kv = ctx.kv("URL_MAPPINGS")?;
-        for link in blocked_links {
-            db::update_link_status_by_id(&db, &link.id, "blocked").await?;
-            if let Err(e) = db::resolve_reports_for_link(
-                &db,
-                &link.id,
-                "reviewed",
-                &format!("Action taken: Blocked {} ({})", match_type, destination),
-                &user_ctx.user_id,
-            )
-            .await
-            {
-                console_log!(
-                    "{}",
-                    serde_json::json!({
-                        "event": "reports_resolve_failed",
-                        "link_id": link.id,
-                        "error": e.to_string(),
-                        "level": "error"
-                    })
-                );
-            }
-
-            match kv::delete_link_mapping(&kv, &link.org_id, &link.short_code).await {
-                Ok(()) => {}
-                Err(e) => {
-                    console_log!(
-                        "{}",
-                        serde_json::json!({
-                            "event": "admin_block_destination_kv_delete_failed",
-                            "short_code": link.short_code,
-                            "link_id": link.id,
-                            "error": e.to_string(),
-                            "level": "error"
-                        })
-                    );
-                }
-            }
-        }
-    }
-
-    Response::from_json(&serde_json::json!({
-        "success": true,
-        "message": "Destination blocked successfully",
-        "blocked_links": blocked_count
-    }))
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/admin/blacklist",
-    tag = "Admin",
-    summary = "List blacklisted destinations",
-    responses(
-        (status = 200, description = "Array of blacklist entries"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_get_blacklist(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    let entries = db::get_all_blacklist(&db).await?;
-
-    Response::from_json(&entries)
-}
-
-#[utoipa::path(
-    delete,
-    path = "/api/admin/blacklist/{id}",
-    tag = "Admin",
-    summary = "Remove blacklist entry",
-    params(("id" = String, Path, description = "Blacklist entry ID")),
-    responses(
-        (status = 200, description = "Entry removed"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_remove_blacklist(
-    req: Request,
-    ctx: RouteContext<()>,
-) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing blacklist entry ID".to_string()))?
-        .to_string();
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    db::remove_from_blacklist(&db, &id).await?;
-
-    Response::from_json(&serde_json::json!({
-        "success": true,
-        "message": "Blacklist entry removed successfully"
-    }))
-}
+// Blacklist handlers moved to api/admin/blacklist.rs
 
 #[utoipa::path(
     put,
